% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crps_2pnorm.R
\name{crps_2pnorm}
\alias{crps_2pnorm}
\title{Calculating the CRPS for the two-piece-normal distribution}
\usage{
crps_2pnorm(y, scale1, scale2, location = 0)
}
\arguments{
\item{y}{vector of observations.}

\item{scale1, scale2}{vectors of positive scale parameters.}

\item{location}{vector of location parameters.}
}
\value{
A vector of CRPS values.
}
\description{
Calculating the CRPS for the two-piece-normal distribution
}
