% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crps_beta.R
\name{crps_beta}
\alias{crps_beta}
\title{Calculating the CRPS for the beta distribution}
\usage{
crps_beta(y, shape1, shape2, lower = 0, upper = 1)
}
\arguments{
\item{y}{vector of observations.}

\item{shape1, shape2}{vectors of positive shape parameters.}

\item{lower, upper}{vectors of lower and upper limits of the distribution. Must be finite.}
}
\value{
A vector of CRPS values.
}
\description{
Calculating the CRPS for the beta distribution
}
