% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crps_exp.R
\name{crps_exp}
\alias{crps_exp}
\alias{crps_expM}
\title{Calculating the CRPS for the exponential distribution}
\usage{
crps_exp(y, rate = 1)

crps_expM(y, location = 0, scale = 1, mass = 0)
}
\arguments{
\item{y}{vector of observations.}

\item{rate}{vector of rates.}

\item{location}{vector of location parameters.}

\item{scale}{vector of positive scale parameters.}

\item{mass}{vector of point masses in \code{location}.}
}
\value{
A vector of CRPS values.
}
\description{
Calculating the CRPS for the exponential distribution, and the exponential
distribution with location-scale transformation and point mass in
\code{location}.
}
