% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crps_gamma.R
\name{crps_gamma}
\alias{crps_gamma}
\title{Calculating the CRPS for the gamma distribution}
\usage{
crps_gamma(y, shape, rate = 1, scale = 1/rate)
}
\arguments{
\item{y}{vector of observations.}

\item{shape}{vector of positive shape parameters.}

\item{rate}{an alternative way to specify the scale.}

\item{scale}{vector of positive scale parameters.}
}
\value{
A vector of CRPS values.
}
\description{
Calculating the CRPS for the gamma distribution
}
