% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crps_gev.R
\name{crps_gev}
\alias{crps_gev}
\title{Calculating the CRPS for the generalized extreme value distribution}
\usage{
crps_gev(y, shape, location = 0, scale = 1)
}
\arguments{
\item{y}{vector of observations.}

\item{shape}{vector of positive shape parameters.}

\item{location}{vector of location parameters.}

\item{scale}{vector of positive scale parameters.}
}
\value{
A vector of CRPS values.
}
\description{
Calculating the CRPS for the generalized extreme value distribution
}
