% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crps_gpd.R
\name{crps_gpd}
\alias{crps_gpd}
\title{Calculating the CRPS for the generalized Pareto distribution}
\usage{
crps_gpd(y, shape, location = 0, scale = 1, mass = 0)
}
\arguments{
\item{y}{vector of observations.}

\item{shape}{vector of positive shape parameters.}

\item{location}{vector of location parameters.}

\item{scale}{vector of positive scale parameters.}

\item{mass}{vector of point masses in \code{location}.}
}
\value{
A vector of CRPS values.
}
\description{
Calculating the CRPS for the generalized Pareto distribution
}
