% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crps_lapl.R
\name{crps_lapl}
\alias{crps_lapl}
\title{Calculating the CRPS for the Laplace distribution}
\usage{
crps_lapl(y, location = 0, scale = 1)
}
\arguments{
\item{y}{vector of observations.}

\item{location}{vector of location parameters.}

\item{scale}{vector of positive scale parameters.}
}
\value{
A vector of CRPS values.
}
\description{
Calculating the CRPS for the Laplace distribution
}
