% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crps_llapl.R
\name{crps_llapl}
\alias{crps_llapl}
\title{Calculating the CRPS for the log-Laplace distribution}
\usage{
crps_llapl(y, locationlog, scalelog)
}
\arguments{
\item{y}{vector of observations.}

\item{locationlog}{vector of location parameters on the log scale.}

\item{scalelog}{vector of positive scale parameters on the log scale.}
}
\value{
A vector of CRPS values.
}
\description{
Calculating the CRPS for the log-Laplace distribution
}
