% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crps_mixnorm.R
\name{crps_mixnorm}
\alias{crps_mixnorm}
\title{Calculating the CRPS for a mixture of normal distribution.}
\usage{
crps_mixnorm(y, m, s, w, exact = TRUE, rel_tol = 1e-06)
}
\arguments{
\item{y}{vector of observations.}

\item{m}{matrix of mean parameters.}

\item{s}{matrix of scale parameters.}

\item{w}{matrix of weights.}

\item{exact}{if \code{TRUE} calculates the analytical solution; otherwise
numerical integration is used.}

\item{rel_tol}{only used if \code{exact} is \code{FALSE}; relative tolerance
for numerical integration.}
}
\value{
A vector of CRPS values.
}
\description{
Calculating the CRPS for a mixture of normal distribution.
}
