% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crps_nbinom.R
\name{crps_nbinom}
\alias{crps_nbinom}
\title{Calculating the CRPS for the negative binomial distribution}
\usage{
crps_nbinom(y, size, prob)
}
\arguments{
\item{y}{vector of observations.}

\item{size}{target for number of successful trials, or dispersion
    parameter (the shape parameter of the gamma mixing distribution).
    Must be strictly positive, need not be integer.}

\item{prob}{probability of success in each trial. \code{0 < prob <= 1}.}
}
\value{
A vector of CRPS values.
}
\description{
Calculating the CRPS for the negative binomial distribution
}
