% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crps_pois.R
\name{crps_pois}
\alias{crps_pois}
\title{Calculating the CRPS for the Poisson distribution}
\usage{
crps_pois(y, lambda)
}
\arguments{
\item{y}{vector of observations.}

\item{lambda}{vector of (non-negative) means.}
}
\value{
A vector of CRPS values.
}
\description{
Calculating the CRPS for the Poisson distribution
}
