% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crps_unif.R
\name{crps_unif}
\alias{crps_unif}
\title{Calculating the CRPS for the uniform distribution}
\usage{
crps_unif(y, min = 0, max = 1, lmass = 0, umass = 0)
}
\arguments{
\item{y}{vector of observations.}

\item{min, max}{lower and upper limits of the distribution. Must be finite.}

\item{lmass, umass}{vectors of point masses in \code{min} and \code{max}
respectively.}
}
\value{
A vector of CRPS values.
}
\description{
Calculating the CRPS for the uniform distribution
}
