\name{sRGev}
\alias{sRGev}
\alias{fgev}
\title{
The Generalized Extreme Value Distribution
}
\description{Density function for the generalized extreme value distribution.}
\usage{
fgev(x, location, scale, shape)
}
\arguments{
  \item{x}{vector of quantiles}
  \item{location}{vector of location parameters}
  \item{scale}{vector of scale parameters}
  \item{shape}{vector of shape parameters}
}
\details{
These functions are implemented for internal use.
For details, see Friederichs, F. and T.L. Thorarinsdottir (2012, "Forecast verification for extreme value distributions with an application to probabilistic peak wind prediction", Environmetrics 23, 579-594).
}
