\name{sRLaplace}
\alias{sRLaplace}
\alias{flapl}
\alias{f2plapl}
\title{
The Laplace Distribution
}
\description{Density function for the double-exponential distribution (Laplace distribution).}
\usage{
flapl(x, location, scale)

f2plapl(x, location, scale1, scale2)
}
\arguments{
  \item{x}{vector of quantiles}
  \item{location}{vector of location parameters}
  \item{scale, scale1, scale2}{vector of scale parameters}
}
\details{
These functions are implemented for internal use.
}
