\name{sRLoglaplace}
\alias{sRLoglaplace}
\alias{fllapl}
\title{
The Log Laplace Distribution
}
\description{Density function for the log Laplace distribution.}
\usage{
fllapl(x, locationlog, scalelog)
}
\arguments{
  \item{x}{vector of quantiles}
  \item{locationlog}{vector of location parameters on the log scale}
  \item{scalelog}{vector of scale parameters on the log scale}
}
\details{
These functions are implemented for internal use.
}
