\name{sRLoglogistic}
\alias{sRLoglogistic}
\alias{fllogis}
\title{
The Log Logistic Distribution
}
\description{Density function for the log logistic distribution.}
\usage{
fllogis(x, locationlog, scalelog)
}
\arguments{
  \item{x}{vector of quantiles}
  \item{locationlog}{vector of location parameters on the log scale}
  \item{scalelog}{vector of scale parameters on the log scale}
}
\details{
These functions are implemented for internal use.
}
