% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_replication.R
\name{plot.casestudy}
\alias{plot.casestudy}
\title{Plot the output of run_casestudy}
\usage{
\method{plot}{casestudy}(x, ...)
}
\arguments{
\item{x}{object of class \code{casestudy}, generated by \link{run_casestudy}}

\item{...}{additional parameters, see details below.}
}
\value{
none, used for the effect of drawing a plot.
}
\description{
Plot the output of run_casestudy
}
\details{
The plot is in the same format as Figure 5 in KLTG (2019). Its content (nr of MCMC chains, 
maximal sample size, etc) depends on the parameters used to generate \link{run_casestudy}. In terms of 
additional inputs (\code{...}), the following are currently implemented:
\itemize{\item \code{scoring_rule}, the scoring rule for which results are to be plotted, 
either \code{"crps"} or \code{"logs"}. Defaults to \code{"crps"}. \item \code{add_main_title}, logical,
whether to add main title to plot. Defaults to \code{TRUE}.}
}
\references{
Krueger, F., Lerch, S., Thorarinsdottir, T.L. and T. Gneiting (2019): `Predictive inference based on Markov chain Monte Carlo output', working paper, Heidelberg Institute for Theoretical Studies, available at \url{http://arxiv.org/abs/1608.06802}.
}
\seealso{
\link{run_casestudy} produces the forecast results summarized by \link{plot.casestudy}
}
\author{
Fabian Krueger
}
\keyword{replication}
