% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_replication.R
\name{plot.mcstudy}
\alias{plot.mcstudy}
\title{Plot the output of run_mcstudy}
\usage{
\method{plot}{mcstudy}(x, ...)
}
\arguments{
\item{x}{object of class \code{mcstudy}, generated by \link{run_mcstudy}}

\item{...}{additional parameters, see details below.}
}
\value{
none, used for the effect of drawing a plot.
}
\description{
Plot the output of run_mcstudy
}
\details{
The plot is in the same format as Figure 1 or 2 in KLTG (2016), depending on the 
parameters set when running \link{run_mcstudy}. These parameters also determine the plot content 
(nr of MCMC chains, maximal sample size, etc). In terms of 
additional inputs (\code{...}), the following are currently implemented:
\itemize{\item \code{scoring_rule}, the scoring rule for which results are to be plotted, 
either \code{"crps"} or \code{"logs"}. Defaults to \code{"crps"}. \item \code{add_main_title}, logical,
whether to add main title to plot. Defaults to \code{TRUE}.}
}
\references{
Krueger, F., Lerch, S., Thorarinsdottir, T.L. and T. Gneiting (2016): `Probabilistic forecasting and comparative model assessment based on Markov chain Monte Carlo output', working paper, Heidelberg Institute for Theoretical Studies, available at \url{http://arxiv.org/abs/1608.06802}.
}
\seealso{
\link{run_mcstudy} produces the simulation results summarized by \link{plot.mcstudy}
}
\author{
Fabian Krueger
}
\keyword{replication}
