% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_replication.R
\name{print.casestudy}
\alias{print.casestudy}
\title{Simple print method for object of class casestudy}
\usage{
\method{print}{casestudy}(x, ...)
}
\arguments{
\item{x}{Object of class casestudy, generated via \link{run_casestudy}}

\item{...}{Additional specifications (presently not in use)}
}
\description{
Simple print method for object of class casestudy
}
