% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_replication.R
\name{print.mcstudy}
\alias{print.mcstudy}
\title{Simple print function for object of class mcstudy}
\usage{
\method{print}{mcstudy}(x, ...)
}
\arguments{
\item{x}{Object of class mcstudy, generated via \link{run_mcstudy}}

\item{...}{Additional specifications (presently not in use)}
}
\description{
Simple print function for object of class mcstudy
}
