% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{scores}
\alias{scores}
\alias{crps}
\alias{logs}
\title{Generic Scoring Rule Calculation}
\usage{
crps(y, ...)

logs(y, ...)
}
\arguments{
\item{y}{an object for which the score is to be calculated}

\item{...}{further arguments passed to or from other methods}
}
\value{
Returns a vector of scores. One for each forecast-observation pair.
}
\description{
Generic functions for calculating the Continuous Ranked Probability Score
and the Logarithmic Score of R objects.

\code{scoringRules} provides default methods
(\link{crps.numeric}, \link{logs.numeric}) to calculate scores of forecasts
that are members of families of parametric distributions.
}
\seealso{
\link{crps.numeric}, \link{logs.numeric}
}
