% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_2pexp.R
\name{scores_2pexp}
\alias{scores_2pexp}
\alias{crps_2pexp}
\alias{logs_2pexp}
\title{Calculating scores for the two-piece-exponential distribution}
\usage{
crps_2pexp(y, scale1, scale2, location = 0)

logs_2pexp(y, scale1, scale2, location = 0)
}
\arguments{
\item{y}{vector of observations.}

\item{scale1, scale2}{vectors of positive scale parameters.}

\item{location}{vector of location parameters.}
}
\value{
A vector of score values.
}
\description{
Calculating scores for the two-piece-exponential distribution
}
