% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_2pnorm.R
\name{scores_2pnorm}
\alias{scores_2pnorm}
\alias{crps_2pnorm}
\alias{logs_2pnorm}
\title{Calculating scores for the two-piece-normal distribution}
\usage{
crps_2pnorm(y, scale1, scale2, location = 0)

logs_2pnorm(y, scale1, scale2, location = 0)
}
\arguments{
\item{y}{vector of observations.}

\item{scale1, scale2}{vectors of positive scale parameters.}

\item{location}{vector of location parameters.}
}
\value{
A vector of score values.
}
\description{
Calculating scores for the two-piece-normal distribution
}
