% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_beta.R
\name{scores_beta}
\alias{scores_beta}
\alias{crps_beta}
\alias{logs_beta}
\title{Calculating scores for the beta distribution}
\usage{
crps_beta(y, shape1, shape2, lower = 0, upper = 1)

logs_beta(y, shape1, shape2, lower = 0, upper = 1)
}
\arguments{
\item{y}{vector of observations.}

\item{shape1, shape2}{vectors of positive shape parameters.}

\item{lower, upper}{vectors of lower and upper limits of the distribution. Must be finite.}
}
\value{
A vector of score values.
}
\description{
Calculating scores for the beta distribution
}
