% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_binom.R
\name{scores_binom}
\alias{scores_binom}
\alias{crps_binom}
\alias{logs_binom}
\title{Calculating scores for the binomial distribution}
\usage{
crps_binom(y, size, prob)

logs_binom(y, size, prob)
}
\arguments{
\item{y}{vector of observations.}

\item{size}{number of trials (zero or more).}

\item{prob}{probability of success on each trial.}
}
\value{
A vector of score values.
}
\description{
Calculating scores for the binomial distribution
}
