% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_exp.R
\name{scores_exp}
\alias{scores_exp}
\alias{crps_exp}
\alias{crps_expM}
\alias{logs_exp}
\alias{logs_exp2}
\alias{dss_exp}
\title{Calculating scores for the exponential distribution}
\usage{
crps_exp(y, rate = 1)

crps_expM(y, location = 0, scale = 1, mass = 0)

logs_exp(y, rate = 1)

logs_exp2(y, location = 0, scale = 1)

dss_exp(y, rate = 1)
}
\arguments{
\item{y}{vector of observations.}

\item{rate}{vector of rates.}

\item{location}{vector of location parameters.}

\item{scale}{vector of positive scale parameters.}

\item{mass}{vector of point masses in \code{location}.}
}
\value{
A vector of score values.
}
\description{
Calculating scores (CRPS, LogS, DSS) for the exponential distribution,
and the exponential distribution with location-scale transformation and
point mass in \code{location}.
}
