% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_gamma.R
\name{scores_gamma}
\alias{scores_gamma}
\alias{crps_gamma}
\alias{logs_gamma}
\alias{dss_gamma}
\title{Calculating scores for the gamma distribution}
\usage{
crps_gamma(y, shape, rate = 1, scale = 1/rate)

logs_gamma(y, shape, rate = 1, scale = 1/rate)

dss_gamma(y, shape, rate = 1, scale = 1/rate)
}
\arguments{
\item{y}{vector of observations.}

\item{shape}{vector of positive shape parameters.}

\item{rate}{an alternative way to specify the scale.}

\item{scale}{vector of positive scale parameters.}
}
\value{
A vector of score values.
}
\description{
Calculating scores for the gamma distribution
}
