% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_gpd.R
\name{scores_gpd}
\alias{scores_gpd}
\alias{crps_gpd}
\alias{logs_gpd}
\title{Calculating scores for the generalized Pareto distribution}
\usage{
crps_gpd(y, shape, location = 0, scale = 1, mass = 0)

logs_gpd(y, shape, location = 0, scale = 1)
}
\arguments{
\item{y}{vector of observations.}

\item{shape}{vector of positive shape parameters.}

\item{location}{vector of location parameters.}

\item{scale}{vector of positive scale parameters.}

\item{mass}{vector of point masses in \code{location}.}
}
\value{
A vector of score values.
}
\description{
Calculating scores for the generalized Pareto distribution
}
