% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_lapl.R
\name{scores_lapl}
\alias{scores_lapl}
\alias{crps_lapl}
\alias{logs_lapl}
\title{Calculating scores for the Laplace distribution}
\usage{
crps_lapl(y, location = 0, scale = 1)

logs_lapl(y, location = 0, scale = 1)
}
\arguments{
\item{y}{vector of observations.}

\item{location}{vector of location parameters.}

\item{scale}{vector of positive scale parameters.}
}
\value{
A vector of score values.
}
\description{
Calculating scores for the Laplace distribution
}
