% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_llapl.R
\name{scores_llapl}
\alias{scores_llapl}
\alias{crps_llapl}
\alias{logs_llapl}
\title{Calculating scores for the log-Laplace distribution}
\usage{
crps_llapl(y, locationlog, scalelog)

logs_llapl(y, locationlog, scalelog)
}
\arguments{
\item{y}{vector of observations.}

\item{locationlog}{vector of location parameters on the log scale.}

\item{scalelog}{vector of positive scale parameters on the log scale.}
}
\value{
A vector of score values.
}
\description{
Calculating scores for the log-Laplace distribution
}
