% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_llogis.R
\name{scores_llogis}
\alias{scores_llogis}
\alias{crps_llogis}
\alias{logs_llogis}
\alias{dss_llogis}
\title{Calculating scores for the log-logistic distribution}
\usage{
crps_llogis(y, locationlog, scalelog)

logs_llogis(y, locationlog, scalelog)

dss_llogis(y, locationlog, scalelog)
}
\arguments{
\item{y}{vector of observations.}

\item{locationlog}{vector of location parameters on the log scale.}

\item{scalelog}{vector of positive scale parameters on the log scale.}
}
\value{
A vector of score values.
}
\description{
Calculating scores for the log-logistic distribution
}
