% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_lnorm.R
\name{scores_lnorm}
\alias{scores_lnorm}
\alias{crps_lnorm}
\alias{logs_lnorm}
\alias{dss_lnorm}
\title{Calculating scores for the log-normal distribution}
\usage{
crps_lnorm(y, meanlog = 0, sdlog = 1, locationlog = meanlog, scalelog = sdlog)

logs_lnorm(y, meanlog = 0, sdlog = 1, locationlog = meanlog, scalelog = sdlog)

dss_lnorm(y, meanlog = 0, sdlog = 1, locationlog = meanlog, scalelog = sdlog)
}
\arguments{
\item{y}{vector of observations.}

\item{meanlog}{an alternative way to specify \code{locationlog}.}

\item{sdlog}{an alternative way to specify \code{scalelog}.}

\item{locationlog}{vector of location parameters on the log scale.}

\item{scalelog}{vector of positive scale parameters on the log scale.}
}
\value{
A vector of score values.
}
\description{
Calculating scores for the log-normal distribution
}
