% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_pois.R
\name{scores_pois}
\alias{scores_pois}
\alias{crps_pois}
\alias{logs_pois}
\alias{dss_pois}
\title{Calculating scores for the Poisson distribution}
\usage{
crps_pois(y, lambda)

logs_pois(y, lambda)

dss_pois(y, lambda)
}
\arguments{
\item{y}{vector of observations.}

\item{lambda}{vector of (non-negative) means.}
}
\value{
A vector of score values.
}
\description{
Calculating scores for the Poisson distribution
}
