% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_sample_multiv_weighted.R
\name{scores_sample_multiv_weighted}
\alias{scores_sample_multiv_weighted}
\alias{twes_sample}
\alias{owes_sample}
\alias{twmmds_sample}
\alias{owmmds_sample}
\alias{twvs_sample}
\alias{owvs_sample}
\title{Weighted Multivariate Scoring Rules for Simulated Forecast Distributions}
\usage{
twes_sample(y, dat, a = -Inf, b = Inf, w = NULL)

owes_sample(y, dat, a = -Inf, b = Inf, w = NULL)

twmmds_sample(y, dat, a = -Inf, b = Inf, w = NULL)

owmmds_sample(y, dat, a = -Inf, b = Inf, w = NULL)

twvs_sample(y, dat, a = -Inf, b = Inf, w_vs = NULL, p = 0.5)

owvs_sample(y, dat, a = -Inf, b = Inf, w = NULL, w_vs = NULL, p = 0.5)
}
\arguments{
\item{y}{realized values (numeric vector of length d).}

\item{dat}{numeric matrix of data
(columns are simulation draws from multivariate forecast distribution).}

\item{a}{numeric vector of of length d containing lower bounds for the indicator 
weight function \code{w(z) = 1{a_{1} < z_{1} < b_{1}, ..., a_{d} < z_{d} < b_{d}}}.}

\item{b}{numeric vector of of length d containing lower bounds for the indicator 
weight function \code{w(z) = 1{a_{1} < z_{1} < b_{1}, ..., a_{d} < z_{d} < b_{d}}}.}

\item{w}{numeric vector of weights for forecast draws (length equal to number of columns of \code{dat})}

\item{w_vs}{numeric matrix of weights for \code{dat} used in the variogram
score. This matrix must be square and symmetric, with all elements being non-negative.
If no weights are specified, constant weights (with all elements of \code{w_vs} 
equal to one) are used.}

\item{p}{order of variogram score. Standard choices include \eqn{p = 1} and
\eqn{p = 0.5}.}
}
\value{
Value of the score. \emph{A lower score indicates a better forecast.}
}
\description{
Compute weighted versions of multivariate scores \eqn{S(y, dat)}, where \eqn{S} is a
proper scoring rule, \eqn{y} is a d-dimensional realization vector and 
\eqn{dat} is a simulated sample of multivariate forecasts. Three scores
are available: The energy score, a score based on a Gaussian kernel 
(\link{mmds_sample}, see details below) and the variogram score of order \eqn{p}.
}
\details{
In the input matrix \code{dat} each column is expected to represent a sample
from the multivariate forecast distribution, the number of rows of \code{dat}
thus has to match the length of the observation vector \code{y}, and the
number of columns of \code{dat} is the number of simulated samples.
}
\references{
\emph{Threshold-weighted scores}

Allen, S., Ginsbourger, D. and J. Ziegel (2022):
`Evaluating forecasts for high-impact events using transformed kernel scores',
arXiv preprint arXiv:2202.12732. 
\doi{10.48550/arXiv.2202.12732}

\emph{Outcome-weighted scores:}
 
Holzmann, H. and B. Klar (2017):
`Focusing on regions of interest in forecast evaluation',
\emph{Annals of Applied Statistics} 11, 2404-2431. 
\doi{10.1214/17-AOAS1088}
}
\author{
Sam Allen
}
