% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_sample_univ.R
\name{scores_sample_univ}
\alias{scores_sample_univ}
\alias{crps_sample}
\alias{logs_sample}
\title{Scoring Rules for Simulated Forecast Distributions}
\usage{
crps_sample(y, dat, method = "edf", w = NULL, bw = NULL,
  num_int = FALSE, show_messages = TRUE)

logs_sample(y, dat, bw = NULL, show_messages = TRUE)
}
\arguments{
\item{y}{realized value (numeric vector of length one).}

\item{dat}{Numeric vector of data
(simulation draws from forecast distribution).}

\item{method}{String, approximation method used in \link{crps_sample}.
Currently implemented: "edf" (empirical distribution function) and
"kde" (kernel density estimation).}

\item{w}{Numeric vector of weights for \code{dat} (if \code{method = "edf"}).}

\item{bw}{Numeric scalar for bandwidth (if \code{method = "kde"}).}

\item{num_int}{Logical; if TRUE numerical integration is used
(if \code{method = "kde"}).}

\item{show_messages}{Logical; on/off switch for messages (does not affect
warnings and error, which are always printed).}
}
\value{
Value of the score. \emph{A lower score indicates a better forecast.}
}
\description{
Compute scores of the form \eqn{S(y, dat)}, where \eqn{S} is a proper
scoring rule, \eqn{y} is a scalar realization and \eqn{dat} is a simulated
sample of data.
}
\details{
When using \code{method = "edf"}, \link{crps_sample} employs an empirical
version of the quantile decomposition of the CRPS (Laio and Tamea, 2007).
When using \code{method = "kde"}, the function does kernel density
estimation using a Gaussian kernel. The bandwidth (\code{bw}) can be
specified manually, in which case it must be a positive number. If
\code{bandwidth == NULL}, the bandwidth is selected using the core function
\link{bw.nrd}. Numerical integration may speed up computation for
\link{crps_sample} in case of large samples \code{dat} if
\code{method = "kde"}.
}
\examples{
sample <- rnorm(1e4)
obs <- 1
crps_sample(y = obs, dat = sample)  
logs_sample(y = obs, dat = sample)

## exact solutions using analytical closed form expression:
crps(y = obs, family = "normal", mean = 0, sd = 1)
logs(y = obs, family = "normal", mean = 0, sd = 1)

}
\references{
\emph{Evaluating simulation based forecast distributions:}

Krueger, F., S. Lerch, T.L. Thorarinsdottir and T. Gneiting (2017),
`Probabilistic forecasting and comparative model assessment based on
Markov Chain Monte Carlo output', working paper,
Heidelberg Institute for Theoretical Studies,
available at \url{http://arxiv.org/abs/1608.06802}.
 
\emph{Empirical quantile decomposition of the CRPS:}
 
Laio, F. and S. Tamea (2007),
`Verification tools for probabilistic forecasts of continuous
hydrological variables',
Hydrology and Earth System Sciences, 11, 1267-1277.
}
\author{
Alexander Jordan, Fabian Krueger
}
