% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_sample_univ_weighted.R
\name{scores_sample_univ_weighted}
\alias{scores_sample_univ_weighted}
\alias{twcrps_sample}
\alias{owcrps_sample}
\alias{clogs_sample}
\title{Weighted Scoring Rules for Simulated Forecast Distributions}
\usage{
twcrps_sample(
  y,
  dat,
  a = -Inf,
  b = Inf,
  method = "edf",
  w = NULL,
  bw = NULL,
  num_int = FALSE,
  show_messages = TRUE
)

owcrps_sample(
  y,
  dat,
  a = -Inf,
  b = Inf,
  method = "edf",
  w = NULL,
  bw = NULL,
  num_int = FALSE,
  show_messages = TRUE,
  comp = TRUE
)

clogs_sample(
  y,
  dat,
  a = -Inf,
  b = Inf,
  w = NULL,
  bw = NULL,
  show_messages = TRUE,
  comp = TRUE
)
}
\arguments{
\item{y}{vector of realized values.}

\item{dat}{vector or matrix (depending on \code{y}; see details)
of simulation draws from forecast distribution.}

\item{a}{optional; numeric lower bound for indicator weight function \code{w(z) = 1{a < z < b}}.}

\item{b}{optional; numeric upper bound for indicator weight function \code{w(z) = 1{a < z < b}}.}

\item{method}{string; approximation method. Options:
"edf" (empirical distribution function) and
"kde" (kernel density estimation).}

\item{w}{optional; vector or matrix (matching \code{dat}) of weights for method \code{"edf"}.}

\item{bw}{optional; vector (matching \code{y}) of bandwidths for kernel density
estimation; see details.}

\item{num_int}{logical; if TRUE numerical integration is used for method \code{"kde"}.}

\item{show_messages}{logical; display of messages (does not affect
warnings and errors).}
}
\value{
Value of the score. \emph{A lower score indicates a better forecast.}
}
\description{
Calculate weighted scores (threshold-weighted CRPS, outcome-weighted CRPS, censored likelihood score) given observations and draws from the predictive distributions.
}
\details{
For a vector \code{y} of length n, \code{dat} should be given as a matrix
with n rows. If \code{y} has length 1, then \code{dat} may be a vector.

\code{\link{twcrps_sample}} transforms \code{y} and \code{dat} according to the
weight function, and then employs \code{\link{crps_sample}}. See the documentation
for \code{\link{crps_sample}} for further details.

The \code{w} argument is used to weight the draws from the predictive distribution. 
This argument is also present in the unweighted scores (e.g. \code{\link{crps_sample}}).
This does not weight particular outcomes within the weighted scoring rules.
}
\references{
\emph{Threshold-weighted CRPS:}

Gneiting, T. and R. Ranjan (2011): 
`Comparing density forecasts using threshold-and quantile-weighted scoring rules', 
\emph{Journal of Business & Economic Statistics} 29, 411-422. 
\doi{10.1198/jbes.2010.08110}
 
\emph{Outcome-weighted CRPS:}
 
Holzmann, H. and B. Klar (2017):
`Focusing on regions of interest in forecast evaluation',
\emph{Annals of Applied Statistics} 11, 2404-2431. 
\doi{10.1214/17-AOAS1088}

\emph{Censored likelihood score:}

Diks, C., Panchenko, V. and D. Van Dijk (2011):
`Likelihood-based scoring rules for comparing density forecasts in tails',
\emph{Journal of Econometrics} 163, 215-230.
\doi{10.1016/j.jeconom.2011.04.001}
}
\author{
Sam Allen
}
