% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_unif.R
\name{scores_unif}
\alias{scores_unif}
\alias{crps_unif}
\alias{logs_unif}
\alias{dss_unif}
\title{Calculating scores for the uniform distribution}
\usage{
crps_unif(y, min = 0, max = 1, lmass = 0, umass = 0)

logs_unif(y, min = 0, max = 1)

dss_unif(y, min = 0, max = 1)
}
\arguments{
\item{y}{vector of observations.}

\item{min, max}{lower and upper limits of the distribution. Must be finite.}

\item{lmass, umass}{vectors of point masses in \code{min} and \code{max}
respectively.}
}
\value{
A vector of score values.
}
\description{
Calculating scores for the uniform distribution
}
