% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_replication.R
\name{summary.casestudy}
\alias{summary.casestudy}
\title{Summary method for class casestudy}
\usage{
\method{summary}{casestudy}(object, ...)
}
\arguments{
\item{object}{Object of class casestudy, generated via \link{run_casestudy}}

\item{...}{Additional specifications (presently not in use)}
}
\description{
Summary method for class casestudy
}
