% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_replication.R
\name{summary.mcstudy}
\alias{summary.mcstudy}
\title{Simple summary method for class mcstudy}
\usage{
\method{summary}{mcstudy}(object, ...)
}
\arguments{
\item{object}{Object of class mcstudy, generated via \link{run_mcstudy}}

\item{...}{Additional specifications (presently not in use)}
}
\description{
Simple summary method for class mcstudy
}
