% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weight_func.R
\name{get_weight_func}
\alias{get_weight_func}
\title{Default Weight and Chaining Functions}
\usage{
get_weight_func(
  name = c("norm_cdf", "norm_surv", "norm_pdf", "logis_cdf", "logis_surv", "logis_pdf"),
  mu = 0,
  sigma = 1,
  weight = TRUE
)
}
\arguments{
\item{name}{name of the weight function to extract.}

\item{mu}{location parameter of the normal or logistic distribution.}

\item{sigma}{scale parameter of the normal or logistic distribution.}

\item{weight}{logical specifying whether to return a weight function (\code{weight = TRUE})
or chaining function (\code{weight = FALSE}).}
}
\value{
A weight or chaining function.
}
\description{
Get commonly used weight or chaining functions to use within weighted scoring rules.
The normal and logistic distribution, density, and survival functions are available.
Multivariate normal distribution functions are also available for multivariate scoring rules.
}
\details{
Details will be added here
}
\examples{
\dontrun{
Examples will be added here
}

}
\references{
Gneiting, T. and R. Ranjan (2011): 
`Comparing density forecasts using threshold-and quantile-weighted scoring rules', 
\emph{Journal of Business & Economic Statistics} 29, 411-422. 
\doi{10.1198/jbes.2010.08110}

Allen, S., Ginsbourger, D. and J. Ziegel (2023): 
`Evaluating forecasts for high-impact events using transformed kernel scores', 
\emph{SIAM/ASA Journal on Uncertainty Quantification} 11, 906-940.
\doi{10.1137/22M1532184}
}
\seealso{
\link{scores_sample_univ_weighted} and \link{scores_sample_multiv_weighted} for weighted scoring rules.
}
\author{
Sam Allen
}
