% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_mixnorm.R
\name{scores_mixnorm}
\alias{scores_mixnorm}
\alias{crps_mixnorm}
\alias{logs_mixnorm}
\title{Calculating scores for a mixture of normal distributions.}
\usage{
crps_mixnorm(y, m, s, w, exact = TRUE, rel_tol = 1e-06)

logs_mixnorm(y, m, s, w)
}
\arguments{
\item{y}{vector of observations.}

\item{m}{matrix of mean parameters.}

\item{s}{matrix of scale parameters.}

\item{w}{matrix of weights.}

\item{exact}{if \code{TRUE} calculates the analytical solution; otherwise
numerical integration is used.}

\item{rel_tol}{only used if \code{exact} is \code{FALSE}; relative tolerance
for numerical integration.}
}
\value{
A vector of score values.
}
\description{
Calculating scores for a mixture of normal distributions.
}
