% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_quantiles.R
\name{scores_quantiles}
\alias{scores_quantiles}
\alias{qs}
\alias{ints}
\title{Quantile and interval scores}
\usage{
qs(y, x, alpha)

ints(y, x_lower, x_upper, target_coverage)
}
\arguments{
\item{y}{vector of observations}

\item{x}{vector of quantile predictions}

\item{alpha}{quantile level of interest}

\item{x_lower, x_upper}{vector of quantile predictions (lower and upper endpoints of prediction intervals)}

\item{target_coverage}{target (i.e., nominal) coverage level of prediction interval}
}
\value{
A vector of score values. Smaller values indicate better forecasts. Note that 
the interval score refers to the central prediction interval at level \code{target_coverage}.
}
\description{
Quantile and interval scores
}
\examples{
# Interval score is proportional to sum of two quantile scores
10*(qs(y = 1, x = qnorm(.1), alpha = .1) + qs(y = 1, x = qnorm(.9), alpha = .9))
ints(y = 1, x_lower = qnorm(.1), x_upper = qnorm(.9), target_coverage = .8)
}
\references{
Quantile score:
Koenker, R. and G. Bassett (1978): `Regression quantiles', Econometrica 46, 33-50. \doi{https://doi.org/10.2307/1913643}
Interval score:
Gneiting, T. and A.E. Raftery (2007):
`Strictly proper scoring rules, prediction and estimation',
Journal of the American Statistical Association 102, 359-378. \doi{10.1198/016214506000001437}
}
